+++ docs/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.5.1] - 2024-04-17
+
+### Added
+
+- !wiki command to search for and return the article summary for a given Wikipedia topic
+- !help command messages for !backmc and !wiki commands
+
+### Fixed
+
+- Minor formatting issue(s) in various !help command messages
+
 ## [5.4.0] - 2024-04-17
 
 ### Added